requirements.txt
requests
python-telegram-bot
PySocks